requirement.txt
absl-py==0.11.0
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1598884132938/work
cycler==0.10.0
cytoolz==0.11.0
dask @ file:///tmp/build/80754af9/dask-core_1610821741412/work
decorator==4.4.2
google-auth==1.24.0
google-auth-oauthlib==0.4.2
grpcio==1.35.0
idna==2.10
imagecodecs @ file:///tmp/build/80754af9/imagecodecs_1611258163564/work
imageio @ file:///tmp/build/80754af9/imageio_1594161405741/work
imgaug @ file:///tmp/build/80754af9/imgaug_1605226543559/work
importlib-metadata==3.4.0
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1604014532738/work
Markdown==3.3.3
matplotlib @ file:///tmp/build/80754af9/matplotlib-base_1603376012865/work
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
networkx @ file:///tmp/build/80754af9/networkx_1598376031484/work
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603479632437/work
oauthlib==3.1.0
olefile==0.46
pandas==1.2.1
Pillow @ file:///tmp/build/80754af9/pillow_1609786740831/work
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
python-dateutil==2.8.1
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyWavelets @ file:///tmp/build/80754af9/pywavelets_1601658308664/work
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scikit-image==0.17.2
scipy @ file:///tmp/build/80754af9/scipy_1597686620742/work
Shapely @ file:///tmp/build/80754af9/shapely_1599817087478/work
six @ file:///tmp/build/80754af9/six_1605205313296/work
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
terminaltables==3.1.0
tifffile @ file:///tmp/build/80754af9/tifffile_1610739638720/work
toolz @ file:///home/linux1/recipes/ci/toolz_1610987900194/work
torch==1.7.1
torchaudio==0.7.0a0+a853dff
torchvision==0.8.2
tornado @ file:///tmp/build/80754af9/tornado_1606942283357/work
tqdm==4.56.0
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
urllib3==1.26.2
Werkzeug==1.0.1
zipp==3.4.0
